TFP-5193: Steg 4, fjerner gammel simuleringslogikk og tilbyr /start endepunkt for fpsak
