+++ domenetjenester/simulering/pom.xml
@@ -24,18 +24,6 @@
             <artifactId>felles-feil</artifactId>
         </dependency>
         <dependency>
-            <groupId>no.nav.foreldrepenger.felles</groupId>
-            <artifactId>felles-xmlutils</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>no.nav.foreldrepenger.oppdrag</groupId>
-            <artifactId>integrasjon</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>no.nav.foreldrepenger.kontrakter</groupId>
-            <artifactId>oppdragskjema-v1</artifactId>
-        </dependency>
-        <dependency>
             <groupId>no.nav.foreldrepenger.oppdrag</groupId>
             <artifactId>domenetjenester-person</artifactId>
         </dependency>
