--- domenetjenester/simulering/src/main/java/no/nav/foreldrepenger/oppdrag/domenetjenester/simulering/fpwsproxy/FeilDto.java
@@ -1,5 +1,0 @@
-package no.nav.foreldrepenger.oppdrag.domenetjenester.simulering.fpwsproxy;
-
-// TODO: Konsolider FeilDto fra web modul. Flytt til felles? kontrakter?
-public record FeilDto(FeilType type, String feilmelding) {
-}
