--- domenetjenester/simulering/src/main/java/no/nav/foreldrepenger/oppdrag/domenetjenester/simulering/fpwsproxy/FeilType.java
@@ -1,9 +1,0 @@
-package no.nav.foreldrepenger.oppdrag.domenetjenester.simulering.fpwsproxy;
-
-public enum FeilType {
-    MANGLER_TILGANG_FEIL,
-    TOMT_RESULTAT_FEIL,
-    OPPDRAG_FORVENTET_NEDETID,
-    GENERELL_FEIL,
-    ;
-}
