+++ domenetjenester/simulering/src/main/java/no/nav/foreldrepenger/oppdrag/domenetjenester/simulering/fpwsproxy/OppdragNedetidException.java
@@ -2,11 +2,11 @@ package no.nav.foreldrepenger.oppdrag.domenetjenester.simulering.fpwsproxy;
 
 import no.nav.vedtak.exception.IntegrasjonException;
 
-public class OppdragNedetidFpWsProxyException extends IntegrasjonException {
+public class OppdragNedetidException extends IntegrasjonException {
 
     private static final String MELDING = "Kallet mot oppdragsystemet feilet. Feilmelding og tidspunktet tilsier at oppdragsystemet har forventet nedetid (utenfor åpningstid).";
 
-    public OppdragNedetidFpWsProxyException() {
+    public OppdragNedetidException() {
         super("FPO-273196", MELDING);
     }
