--- domenetjenester/simulering/src/test/resources/xml/oppdrag_refusjon.xml
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<ns2:oppdrag xmlns:ns2="http://www.trygdeetaten.no/skjema/oppdrag">
-    <oppdrag-110>
-        <kodeAksjon>1</kodeAksjon>
-        <kodeEndring>NY</kodeEndring>
-        <kodeFagomraade>FPREF</kodeFagomraade>
-        <fagsystemId>135702910101</fagsystemId>
-        <utbetFrekvens>MND</utbetFrekvens>
-        <oppdragGjelderId>22067300444</oppdragGjelderId>
-        <datoOppdragGjelderFom>2000-01-01</datoOppdragGjelderFom>
-        <saksbehId>Z991095</saksbehId>
-        <avstemming-115>
-            <kodeKomponent>VLFP</kodeKomponent>
-            <nokkelAvstemming>2017-12-13-10.25.40.744</nokkelAvstemming>
-            <tidspktMelding>2017-12-13-10.25.40.744</tidspktMelding>
-        </avstemming-115>
-        <oppdrags-enhet-120>
-            <typeEnhet>BOS</typeEnhet>
-            <enhet>8020</enhet>
-            <datoEnhetFom>1900-01-01</datoEnhetFom>
-        </oppdrags-enhet-120>
-        <oppdrags-linje-150>
-            <kodeEndringLinje>NY</kodeEndringLinje>
-            <vedtakId>2017-12-13</vedtakId>
-            <delytelseId>135702910101100</delytelseId>
-            <kodeKlassifik>FPREFAG-IOP</kodeKlassifik>
-            <datoVedtakFom>2017-09-07</datoVedtakFom>
-            <datoVedtakTom>2017-09-09</datoVedtakTom>
-            <sats>1177</sats>
-            <fradragTillegg>T</fradragTillegg>
-            <typeSats>DAG</typeSats>
-            <brukKjoreplan>N</brukKjoreplan>
-            <saksbehId>Z991095</saksbehId>
-            <henvisning>999951</henvisning>
-            <refusjonsinfo-156>
-                <maksDato>2017-10-10</maksDato>
-                <refunderesId>00973861778</refunderesId>
-                <datoFom>2017-12-13</datoFom>
-            </refusjonsinfo-156>
-            <grad-170>
-                <typeGrad>UFOR</typeGrad>
-                <grad>100</grad>
-            </grad-170>
-            <attestant-180>
-                <attestantId>Z991095</attestantId>
-            </attestant-180>
-        </oppdrags-linje-150>
-        <oppdrags-linje-150>
-            <kodeEndringLinje>NY</kodeEndringLinje>
-            <vedtakId>2017-12-13</vedtakId>
-            <delytelseId>135702910101101</delytelseId>
-            <kodeKlassifik>FPREFAG-IOP</kodeKlassifik>
-            <datoVedtakFom>2017-09-10</datoVedtakFom>
-            <datoVedtakTom>2017-09-28</datoVedtakTom>
-            <sats>1177</sats>
-            <fradragTillegg>T</fradragTillegg>
-            <typeSats>DAG</typeSats>
-            <brukKjoreplan>N</brukKjoreplan>
-            <saksbehId>Z991095</saksbehId>
-            <henvisning>999951</henvisning>
-            <refFagsystemId>135702910101</refFagsystemId>
-            <refDelytelseId>135702910101100</refDelytelseId>
-            <refusjonsinfo-156>
-                <maksDato>2017-10-10</maksDato>
-                <refunderesId>00973861778</refunderesId>
-                <datoFom>2017-12-13</datoFom>
-            </refusjonsinfo-156>
-            <grad-170>
-                <typeGrad>UFOR</typeGrad>
-                <grad>100</grad>
-            </grad-170>
-            <attestant-180>
-                <attestantId>Z991095</attestantId>
-            </attestant-180>
-        </oppdrags-linje-150>
-        <oppdrags-linje-150>
-            <kodeEndringLinje>NY</kodeEndringLinje>
-            <vedtakId>2017-12-13</vedtakId>
-            <delytelseId>135702910101102</delytelseId>
-            <kodeKlassifik>FPREFAG-IOP</kodeKlassifik>
-            <datoVedtakFom>2017-10-01</datoVedtakFom>
-            <datoVedtakTom>2017-10-10</datoVedtakTom>
-            <sats>1177</sats>
-            <fradragTillegg>T</fradragTillegg>
-            <typeSats>DAG</typeSats>
-            <brukKjoreplan>N</brukKjoreplan>
-            <saksbehId>Z991095</saksbehId>
-            <henvisning>999951</henvisning>
-            <refFagsystemId>135702910101</refFagsystemId>
-            <refDelytelseId>135702910101101</refDelytelseId>
-            <refusjonsinfo-156>
-                <maksDato>2017-10-10</maksDato>
-                <refunderesId>00973861778</refunderesId>
-                <datoFom>2017-12-13</datoFom>
-            </refusjonsinfo-156>
-            <grad-170>
-                <typeGrad>UFOR</typeGrad>
-                <grad>100</grad>
-            </grad-170>
-            <attestant-180>
-                <attestantId>Z991095</attestantId>
-            </attestant-180>
-        </oppdrags-linje-150>
-        <oppdrags-linje-150>
-            <kodeEndringLinje>NY</kodeEndringLinje>
-            <vedtakId>2017-12-13</vedtakId>
-            <delytelseId>135702910101103</delytelseId>
-            <kodeKlassifik>FPREFAGFER-IOP</kodeKlassifik>
-            <datoVedtakFom>2018-05-01</datoVedtakFom>
-            <datoVedtakTom>2018-05-31</datoVedtakTom>
-            <sats>2881</sats>
-            <fradragTillegg>T</fradragTillegg>
-            <typeSats>ENG</typeSats>
-            <brukKjoreplan>N</brukKjoreplan>
-            <saksbehId>Z991095</saksbehId>
-            <henvisning>999951</henvisning>
-            <refusjonsinfo-156>
-                <maksDato>2017-10-10</maksDato>
-                <refunderesId>00973861778</refunderesId>
-                <datoFom>2017-12-13</datoFom>
-            </refusjonsinfo-156>
-            <attestant-180>
-                <attestantId>Z991095</attestantId>
-            </attestant-180>
-        </oppdrags-linje-150>
-    </oppdrag-110>
-</ns2:oppdrag>
