--- integrasjon/src/main/java/no/nav/foreldrepenger/oppdrag/OppdragConsumer.java
@@ -1,10 +1,0 @@
-package no.nav.foreldrepenger.oppdrag;
-
-import no.nav.system.os.tjenester.simulerfpservice.simulerfpservicegrensesnitt.SimulerBeregningRequest;
-import no.nav.system.os.tjenester.simulerfpservice.simulerfpservicegrensesnitt.SimulerBeregningResponse;
-
-public interface OppdragConsumer {
-
-    SimulerBeregningResponse hentSimulerBeregningResponse(SimulerBeregningRequest simulerBeregningRequest);
-
-}
