--- integrasjon/src/main/java/no/nav/foreldrepenger/oppdrag/OppdragNedetidException.java
@@ -1,10 +1,0 @@
-package no.nav.foreldrepenger.oppdrag;
-
-import no.nav.vedtak.exception.IntegrasjonException;
-
-public class OppdragNedetidException extends IntegrasjonException {
-    public OppdragNedetidException(String kode, String melding, Throwable cause) {
-        super(kode, melding, cause);
-    }
-
-}
