--- integrasjon/src/main/java/no/nav/foreldrepenger/oppdrag/OppdragSelftestConsumer.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.oppdrag;
-
-import no.nav.vedtak.felles.integrasjon.felles.ws.SelftestConsumer;
-
-public interface OppdragSelftestConsumer extends SelftestConsumer {
-}
