--- integrasjon/src/test/resources/oppdrag.properties
@@ -1,10 +1,0 @@
-#Default properties for End2End Integrasjontest av grensesnitt
-# NB: Dette bør ligge i jobb for å kjøre E2ETest i stedet for her.
-OpenIdConnect.issoHost=https://isso-q.adeo.no:443/isso
-OpenIdConnect.issoIssuer=https://isso-q.adeo.no:443/isso/oauth2
-OpenIdConnect.issoJwks=https://isso-q.adeo.no:443/isso/oauth2/connect/jwk_uri
-OpenIdConnect.username=fpoppdrag-q10
-#OpenIdConnect.password=*HEMMELIG - SETT SELV UTENFOR DENNE FILEN *
-securityTokenService.url=https://sts-q10.preprod.local/SecurityTokenServiceProvider/
-systembruker.username=srvfpoppdrag
-Oppdrag.service.url=http://ztest.test.local:9281/cics/services/oppdragService
