+++ web/pom.xml
@@ -103,10 +103,6 @@
             <groupId>no.nav.foreldrepenger.kontrakter</groupId>
             <artifactId>fp-ws-proxy-v1</artifactId>
         </dependency>
-        <dependency>
-            <groupId>no.nav.foreldrepenger.kontrakter</groupId>
-            <artifactId>oppdragskjema-v1</artifactId>
-        </dependency>
 
         <!-- CDI -->
         <dependency>
