--- web/src/test/resources/xml/oppdrag_mottaker.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<ns2:oppdrag xmlns:ns2="http://www.trygdeetaten.no/skjema/oppdrag">
-    <oppdrag-110>
-        <kodeAksjon>1</kodeAksjon>
-        <kodeEndring>NY</kodeEndring>
-        <kodeFagomraade>FP</kodeFagomraade>
-        <fagsystemId>130158784100</fagsystemId>
-        <utbetFrekvens>MND</utbetFrekvens>
-        <oppdragGjelderId>15088011020</oppdragGjelderId>
-        <datoOppdragGjelderFom>2000-01-01</datoOppdragGjelderFom>
-        <saksbehId>Z991097</saksbehId>
-        <avstemming-115>
-            <kodeKomponent>VLFP</kodeKomponent>
-            <nokkelAvstemming>2018-08-16-15.36.55.543</nokkelAvstemming>
-            <tidspktMelding>2018-08-16-15.36.55.543</tidspktMelding>
-        </avstemming-115>
-        <oppdrags-enhet-120>
-            <typeEnhet>BOS</typeEnhet>
-            <enhet>8020</enhet>
-            <datoEnhetFom>1900-01-01</datoEnhetFom>
-        </oppdrags-enhet-120>
-        <oppdrags-linje-150>
-            <kodeEndringLinje>NY</kodeEndringLinje>
-            <vedtakId>2018-08-16</vedtakId>
-            <delytelseId>130158784100100</delytelseId>
-            <kodeKlassifik>FPATFRI</kodeKlassifik>
-            <datoVedtakFom>2018-05-11</datoVedtakFom>
-            <datoVedtakTom>2018-05-31</datoVedtakTom>
-            <sats>738</sats>
-            <fradragTillegg>T</fradragTillegg>
-            <typeSats>DAG</typeSats>
-            <brukKjoreplan>N</brukKjoreplan>
-            <saksbehId>Z991097</saksbehId>
-            <utbetalesTilId>15088011020</utbetalesTilId>
-            <henvisning>1000202</henvisning>
-            <grad-170>
-                <typeGrad>UFOR</typeGrad>
-                <grad>100</grad>
-            </grad-170>
-            <attestant-180>
-                <attestantId>Z991097</attestantId>
-            </attestant-180>
-        </oppdrags-linje-150>
-    </oppdrag-110>
-</ns2:oppdrag>
